Module evm_arithmetization::byte_packing::byte_packing_stark

This crate enforces the correctness of reading and writing sequences of bytes in Big-Endian ordering from and to the memory.

The trace layout consists in one row for an N byte sequence (where 32 ≥ N > 0).

At each row the i-th byte flag will be activated to indicate a sequence of length i+1.

The length of a sequence can be retrieved for CTLs as:

sequence_length = \sum_{i=0}^31 b[i] * (i + 1)

where b[i] is the i-th byte flag.

Because of the discrepancy in endianness between the different tables, the byte sequences are actually written in the trace in reverse order from the order they are provided. We only store the virtual address virt of the first byte, and the virtual address for byte i can be recovered as: virt_i = virt + sequence_length - 1 - i

Note that, when writing a sequence of bytes to memory, both the U256 value and the corresponding sequence length are being read from the stack. Because of the endianness discrepancy mentioned above, we first convert the value to a byte sequence in Little-Endian, then resize the sequence to prune unneeded zeros before reverting the sequence order. This means that the higher-order bytes will be thrown away during the process, if the value is greater than 256^length, and as a result a different value will be stored in memory.