Trait evm_arithmetization::extension_tower::Adj

pub trait Adj: Sized {
    const FROB_T: [[Self; 6]; 2];
    const FROB_Z: [Self; 12];

    // Required method
    fn mul_adj(self) -> Self;
}

This trait defines the method which multiplies by the Fp2 element t^3 whose cube root we will adjoin in the subsequent cubic extension. For BN254 this is 9+i, and for BLS381 it is 1+i. It also defines the relevant FROB constants, given by t^(p^n) and (t^2)^(p^n) for various n, used to compute the frobenius operations.

Required Associated Constants

const FROB_T: [[Self; 6]; 2]

const FROB_Z: [Self; 12]

Required Methods

fn mul_adj(self) -> Self

Object Safety

This trait is not object safe.