Module evm_arithmetization::fixed_recursive_verifier

Structs

AllRecursiveCircuits

Contains all recursive circuits used in the system. For each STARK and each initial degree_bits, this contains a chain of recursive circuits for shrinking that STARK from degree_bits to a constant THRESHOLD_DEGREE_BITS. It also contains a special root circuit for combining each STARK’s shrunk wrapper proof into a single proof.

BlockCircuitData

Data for the block circuit, which is used to generate a final block proof, and compress it with an optional parent proof if present.

ProverOutputData

RecursiveCircuitsForTable

A map between initial degree sizes and their associated shrinking recursion circuits.

RecursiveCircuitsForTableSize

A chain of shrinking wrapper circuits, ending with a final circuit with degree_bits THRESHOLD_DEGREE_BITS.

RootCircuitData

Data for the EVM root circuit, which is used to combine each STARK’s shrunk wrapper proof into a single proof.

SegmentAggregationCircuitData

Data for the segment aggregation circuit, which is used to compress two segment proofs into one. Each inner proof can be either an EVM root proof or another segment aggregation proof.

TwoToOneBlockCircuitData

Data for the two-to-one block circuit, which is used to generate a proof of two unrelated proofs.

TxnAggregationCircuitData

Data for the transaction aggregation circuit, which is used to compress two proofs into one. Each inner proof can be either a segment aggregation proof or another transaction aggregation proof.

Functions

extract_block_public_values

Extracts the two-to-one block aggregation public values of the block from a public inputs slice.

extract_two_to_one_block_hash

Extracts the two-to-one block aggregation hash from a public inputs slice.

verification_key_len

Computes the length added to the public inputs vector by [CircuitBuilder::add_verifier_data_public_inputs].