Struct evm_arithmetization::fixed_recursive_verifier::AllRecursiveCircuits

pub struct AllRecursiveCircuits<F, C, const D: usize>
where
    F: RichField + Extendable<D>,
    C: GenericConfig<D, F = F>,
    C::Hasher: AlgebraicHasher<F>,
{
    pub root: RootCircuitData<F, C, D>,
    pub segment_aggregation: SegmentAggregationCircuitData<F, C, D>,
    pub txn_aggregation: TxnAggregationCircuitData<F, C, D>,
    pub block: BlockCircuitData<F, C, D>,
    pub two_to_one_block: TwoToOneBlockCircuitData<F, C, D>,
    pub by_table: [RecursiveCircuitsForTable<F, C, D>; 9],
}

Contains all recursive circuits used in the system. For each STARK and each initial degree_bits, this contains a chain of recursive circuits for shrinking that STARK from degree_bits to a constant THRESHOLD_DEGREE_BITS. It also contains a special root circuit for combining each STARK’s shrunk wrapper proof into a single proof.

Fields
root: RootCircuitData<F, C, D>

The EVM root circuit, which aggregates the (shrunk) per-table recursive proofs.

segment_aggregation: SegmentAggregationCircuitData<F, C, D>

The segment aggregation circuit, which verifies that two segment proofs that can either be root or aggregation proofs.

txn_aggregation: TxnAggregationCircuitData<F, C, D>

The transaction aggregation circuit, which verifies the aggregation of two proofs that can either be a segment aggregation representing a transaction or an aggregation of transactions.

block: BlockCircuitData<F, C, D>

The block circuit, which verifies a transaction aggregation proof and an optional previous block proof.

two_to_one_block: TwoToOneBlockCircuitData<F, C, D>

The two-to-one block aggregation circuit, which verifies two unrelated block proofs.

by_table: [RecursiveCircuitsForTable<F, C, D>; 9]

Holds chains of circuits for each table and for each initial degree_bits.

Implementations

impl<F, C, const D: usize> AllRecursiveCircuits<F, C, D>
where F: RichField + Extendable<D>, C: GenericConfig<D, F = F> + 'static, C::Hasher: AlgebraicHasher<F>,

pub fn to_bytes( &self, skip_tables: bool, gate_serializer: &dyn GateSerializer<F, D>, generator_serializer: &dyn WitnessGeneratorSerializer<F, D>, ) -> IoResult<Vec<u8>>

Serializes all these preprocessed circuits into a sequence of bytes.

Arguments

skip_tables: a boolean indicating whether to serialize only the upper circuits or the entire prover state, including recursive circuits to shrink STARK proofs.

gate_serializer: a custom gate serializer needed to serialize recursive circuits common data.

generator_serializer: a custom generator serializer needed to serialize recursive circuits proving data.

pub fn from_bytes( bytes: &[u8], skip_tables: bool, gate_serializer: &dyn GateSerializer<F, D>, generator_serializer: &dyn WitnessGeneratorSerializer<F, D>, ) -> IoResult<Self>

Deserializes a sequence of bytes into an entire prover state containing all recursive circuits.

Arguments

bytes: a slice of bytes to deserialize this prover state from.

skip_tables: a boolean indicating whether to deserialize only the upper circuits or the entire prover state, including recursive circuits to shrink STARK proofs.

gate_serializer: a custom gate serializer needed to serialize recursive circuits common data.

generator_serializer: a custom generator serializer needed to serialize recursive circuits proving data.

pub fn new( all_stark: &AllStark<F, D>, degree_bits_ranges: &[Range<usize>; 9], stark_config: &StarkConfig, ) -> Self

Preprocess all recursive circuits used by the system.

Arguments

all_stark: a structure defining the logic of all STARK modules and their associated cross-table lookups.

degree_bits_ranges: the logarithmic ranges to be supported for the recursive tables.

Transactions may yield arbitrary trace lengths for each STARK module (within some bounds), unknown prior generating the witness to create a proof. Thus, for each STARK module, we construct a map from 2^{degree_bits} = length to a chain of shrinking recursion circuits, starting from that length, for each degree_bits in the range specified for this STARK module. Specifying a wide enough range allows a prover to cover all possible scenarios.

stark_config: the configuration to be used for the STARK prover. It will usually be a fast one yielding large proofs.

pub fn final_verifier_data(&self) -> VerifierCircuitData<F, C, D>

Outputs the VerifierCircuitData needed to verify any block proof generated by an honest prover. While the AllRecursiveCircuits prover state can also verify proofs, verifiers only need a fraction of the state to verify proofs. This allows much less powerful entities to behave as verifiers, by only loading the necessary data to verify block proofs.

Usage

let prover_state = AllRecursiveCircuits { ... };
let verifier_state = prover_state.final_verifier_data();

// Verify a provided block proof
assert!(verifier_state.verify(&block_proof).is_ok());

pub fn prove_segment( &self, all_stark: &AllStark<F, D>, config: &StarkConfig, generation_inputs: TrimmedGenerationInputs, segment_data: &mut GenerationSegmentData, timing: &mut TimingTree, abort_signal: Option<Arc<AtomicBool>>, ) -> Result<ProverOutputData<F, C, D>>

For a given transaction payload passed as GenerationInputs, create a proof for each STARK module, then recursively shrink and combine them, eventually culminating in a transaction proof, also called root proof.

Arguments

all_stark: a structure defining the logic of all STARK modules and their associated cross-table lookups.

config: the configuration to be used for the STARK prover. It will usually be a fast one yielding large proofs.

generation_inputs: a transaction and auxiliary data needed to generate a proof, provided in Intermediary Representation.

timing: a profiler defining a scope hierarchy and the time consumed by each one.

abort_signal: an optional AtomicBool wrapped behind an Arc, to send a kill signal early. This is only necessary in a distributed setting where a worker may be blocking the entire queue.

Outputs

This method outputs a tuple of [ProofWithPublicInputs<F, C, D>] and its PublicValues. Only the proof with public inputs is necessary for a verifier to assert correctness of the computation, but the public values are output for the prover convenience, as these are necessary during proof aggregation.

pub fn prove_all_segments( &self, all_stark: &AllStark<F, D>, config: &StarkConfig, generation_inputs: GenerationInputs, max_cpu_len_log: usize, timing: &mut TimingTree, abort_signal: Option<Arc<AtomicBool>>, ) -> Result<Vec<ProverOutputData<F, C, D>>>

Returns a proof for each segment that is part of a full transaction proof.

pub fn prove_segment_after_initial_stark( &self, all_proof: AllProof<F, C, D>, table_circuits: &[(RecursiveCircuitsForTableSize<F, C, D>, u8); 9], abort_signal: Option<Arc<AtomicBool>>, ) -> Result<(ProofWithPublicInputs<F, C, D>, PublicValues)>

From an initial set of STARK proofs passed with their associated recursive table circuits, generate a recursive transaction proof. It is aimed at being used when preprocessed table circuits have not been loaded to memory.

Note: The type of the table_circuits passed as arguments is &[(RecursiveCircuitsForTableSize<F, C, D>, u8); NUM_TABLES]. In particular, for each STARK proof contained within the AllProof object provided to this method, we need to pass a tuple of RecursiveCircuitsForTableSize<F, C, D> and a u8. The former is the recursive chain corresponding to the initial degree size of the associated STARK proof. The latter is the index of this degree in the range that was originally passed when constructing the entire prover state.

Usage

// Load a prover state without its recursive table circuits.
let gate_serializer = DefaultGateSerializer;
let generator_serializer = DefaultGeneratorSerializer::<C, D>::new();
let initial_ranges = [16..25, 10..20, 12..25, 14..25, 9..20, 12..20, 17..30];
let prover_state = AllRecursiveCircuits::<F, C, D>::new(
    &all_stark,
    &initial_ranges,
    &config,
);

// Generate a proof from the provided inputs.
let stark_proof = prove::<F, C, D>(&all_stark, &config, inputs, &mut timing, abort_signal).unwrap();

// Read the degrees of the internal STARK proofs.
// Indices to be passed along the recursive tables
// can be easily recovered as `initial_ranges[i]` - `degrees[i]`.
let degrees = proof.degree_bits(&config);

// Retrieve the corresponding recursive table circuits for each table with the corresponding degree.
let table_circuits = { ... };

// Finally shrink the STARK proof.
let (proof, public_values) = prove_segment_after_initial_stark(
    &all_stark,
    &config,
    &stark_proof,
    &table_circuits,
    &mut timing,
    abort_signal,
).unwrap();

pub fn verify_root( &self, agg_proof: ProofWithPublicInputs<F, C, D>, ) -> Result<()>

pub fn prove_segment_aggregation( &self, lhs_is_agg: bool, lhs_prover_output: &ProverOutputData<F, C, D>, rhs_is_agg: bool, rhs_prover_output: &ProverOutputData<F, C, D>, ) -> Result<ProverOutputData<F, C, D>>

Create an aggregation proof, combining two contiguous proofs into a single one. The combined proofs are segment proofs: they are proofs of some parts of one execution.

While regular root proofs can only assert validity of a single segment of a transaction, segment aggregation proofs can cover an arbitrary range, up to an entire transaction.

Arguments

lhs_is_agg: a boolean indicating whether the left child proof is an aggregation proof or a regular segment proof.

lhs_proof: the left child prover output data.

rhs_is_agg: a boolean indicating whether the right child proof is an aggregation proof or a regular transaction proof.

rhs_proof: the right child prover output data.

Outputs

This method outputs a ProverOutputData<F, C, D>. Only the proof with public inputs is necessary for a verifier to assert correctness of the computation, but the public values and is_dummy are output for the prover convenience, as these are necessary during proof aggregation.

pub fn verify_segment_aggregation( &self, agg_proof: &ProofWithPublicInputs<F, C, D>, ) -> Result<()>

pub fn prove_transaction_aggregation( &self, lhs_is_agg: bool, lhs_proof: &ProofWithPublicInputs<F, C, D>, lhs_public_values: PublicValues, rhs_is_agg: bool, rhs_proof: &ProofWithPublicInputs<F, C, D>, rhs_public_values: PublicValues, ) -> Result<(ProofWithPublicInputs<F, C, D>, PublicValues)>

Creates a final transaction proof, once all segments of a given transaction have been combined into a single aggregation proof.

Transaction proofs can either be generated as a standalone, or combined with a previous transaction proof to assert validity of a range of transactions.

Arguments

opt_parent_txn_proof: an optional parent transaction proof. Passing one will generate a proof of validity for both the transaction range covered by the previous proof and the current transaction.

agg_proof: the final aggregation proof containing all segments within the current transaction.

public_values: the public values associated to the aggregation proof.

Outputs

This method outputs a tuple of [ProofWithPublicInputs<F, C, D>] and its PublicValues. Only the proof with public inputs is necessary for a verifier to assert correctness of the computation.

pub fn verify_txn_aggregation( &self, txn_proof: &ProofWithPublicInputs<F, C, D>, ) -> Result<()>

pub fn prove_block( &self, opt_parent_block_proof: Option<&ProofWithPublicInputs<F, C, D>>, agg_root_proof: &ProofWithPublicInputs<F, C, D>, public_values: PublicValues, ) -> Result<(ProofWithPublicInputs<F, C, D>, FinalPublicValues)>

Create a final block proof, once all transactions of a given block have been combined into a single aggregation proof.

Block proofs can either be generated as standalone, or combined with a previous block proof to assert validity of a range of blocks.

Arguments

opt_parent_block_proof: an optional parent block proof. Passing one will generate a proof of validity for both the block range covered by the previous proof and the current block.

agg_root_proof: the final aggregation proof containing all transactions within the current block.

public_values: the public values associated to the aggregation proof.

Outputs

This method outputs a tuple of [ProofWithPublicInputs<F, C, D>] and its PublicValues. Only the proof with public inputs is necessary for a verifier to assert correctness of the computation.

pub fn verify_block( &self, block_proof: &ProofWithPublicInputs<F, C, D>, ) -> Result<()>

pub fn prove_two_to_one_block( &self, lhs: &ProofWithPublicInputs<F, C, D>, lhs_is_agg: bool, rhs: &ProofWithPublicInputs<F, C, D>, rhs_is_agg: bool, ) -> Result<ProofWithPublicInputs<F, C, D>>

Aggregates two proofs in manner similar to [prove_aggregation].

Arguments

lhs: a proof of either a block or previous aggregation.

lhs_is_agg: specify which case lhs was.

rhs: a proof of either a block or previous aggregation.

rhs_is_agg: specify which case rhs was.

Outputs

Returns a [ProofWithPublicInputs<F, C, D>].

pub fn verify_two_to_one_block( &self, proof: &ProofWithPublicInputs<F, C, D>, ) -> Result<()>

Verifies an existing block aggregation proof.

Arguments

proof: The proof generated with prove_two_to_one_block.

Outputs

Returns whether the proof was valid or not.