Module evm_arithmetization::proof

Structs

AllProof

A STARK proof for each table, plus some metadata used to create recursive wrapper proofs.

BlockHashes

User-provided helper values to compute the BLOCKHASH opcode. The proofs across consecutive blocks ensure that these values are consistent (i.e. shifted by one to the left).

BlockHashesTarget

Circuit version of BlockHashes. Targets for the user-provided previous 256 block hashes and current block hash. Each block hash requires 8 Targets. The proofs across consecutive blocks ensure that these values are consistent (i.e. shifted by eight Targets to the left).

BlockMetadata

Metadata contained in a block header. Those are identical between all state transition proofs within the same block.

BlockMetadataTarget

Circuit version of BlockMetadata. Metadata contained in a block header. Those are identical between all state transition proofs within the same block.

ExtraBlockData

Additional block data that are specific to the local transaction being proven, unlike BlockMetadata.

ExtraBlockDataTarget

Circuit version of ExtraBlockData. Additional block data that are specific to the local transaction being proven, unlike BlockMetadata.

FinalPublicValues

Memory values which are public.

MemCap

Structure for a Merkle cap. It is used for MemBefore and MemAfter.

MemCapTarget

PublicValues

Memory values which are public.

PublicValuesTarget

Memory values which are public. Note: All the larger integers are encoded with 32-bit limbs in little-endian order.

RegistersData

Registers data used to preinitialize the registers and check the final registers of the current proof.

RegistersDataTarget

Circuit version of RegistersData. Registers data used to preinitialize the registers and check the final registers of the current proof.

TrieRoots

Trie hashes.

TrieRootsTarget

Circuit version of TrieRoots. Targets for trie hashes. Since a Target holds a 32-bit limb, each hash requires 8 Targets.