Module evm_arithmetization::testing_utils

A set of utility functions and constants to be used by evm_arithmetization unit and integration tests.

Constants

BEACON_ROOTS_ACCOUNT

BEACON_ROOTS_CONTRACT_ADDRESS

BEACON_ROOTS_CONTRACT_ADDRESS_HASHED

BEACON_ROOTS_CONTRACT_CODE

BEACON_ROOTS_CONTRACT_CODE_HASH

BEACON_ROOTS_CONTRACT_STATE_KEY

BLOB_BASE_FEE_UPDATE_FRACTION

EMPTY_NODE_HASH

G2_TRUSTED_SETUP_POINT

GLOBAL_EXIT_ROOT_ACCOUNT

GLOBAL_EXIT_ROOT_ADDRESS_HASHED

GLOBAL_EXIT_ROOT_STORAGE_POS

Taken from https://github.com/0xPolygonHermez/cdk-erigon/blob/61f0b6912055c73f6879ea7e9b5bac22ea5fc85c/zk/utils/global_exit_root.go#L17.

HISTORY_BUFFER_LENGTH

KZG_VERSIONED_HASH

MIN_BASE_FEE_PER_BLOB_GAS

POINT_EVALUATION_PRECOMPILE_RETURN_VALUE

Functions

beacon_roots_account_nibbles

Returns the Nibbles corresponding to the beacon roots contract account.

beacon_roots_contract_from_storage

Returns the beacon roots contract account from its provided storage trie.

create_account_storage

Creates a storage trie for an account, given a list of (slot, value) pairs.

eth_to_wei

Converts an amount in ETH to wei units.

ger_account_nibbles

Returns the Nibbles corresponding to the beacon roots contract account.

ger_contract_from_storage

init_logger

preinitialized_state_and_storage_tries

Returns an initial state trie containing the beacon roots and global exit roots contracts, along with their storage tries.

sd2u

Converts a decimal string to a U256.

sh2u

Converts an hexadecimal string to a U256.

update_beacon_roots_account_storage

Updates the beacon roots account storage with the provided timestamp and block parent root.

update_ger_account_storage