Type Alias evm_arithmetization::Node

pub type Node = Node<HashedPartialTrie>;

Aliased Type

enum Node {
    Empty,
    Hash(H256),
    Branch {
        children: [Arc<Box<HashedPartialTrie>>; 16],
        value: Vec<u8>,
    },
    Extension {
        nibbles: Nibbles,
        child: Arc<Box<HashedPartialTrie>>,
    },
    Leaf {
        nibbles: Nibbles,
        value: Vec<u8>,
    },
}

Variants

Empty

An empty trie.

Hash(H256)

The digest of trie whose data does not need to be stored.

Important note: Hash nodes should only be created to replace PartialTries whose RLP encoding is >= 32 bytes. Creating a hash node for a PartialTrie smaller than this will cause an incorrect hash to be generated for the trie.

Branch

A branch node, which consists of 16 children and an optional value.

Fields

children: [Arc<Box<HashedPartialTrie>>; 16]

A slice containing the 16 children of this branch node.

value: Vec<u8>

The payload of this node.

Extension

An extension node, which consists of a list of nibbles and a single child.

Fields

nibbles: Nibbles

The path of this extension.

child: Arc<Box<HashedPartialTrie>>

The child of this extension node.

Leaf

A leaf node, which consists of a list of nibbles and a value.

Fields

nibbles: Nibbles

The path of this leaf node.

value: Vec<u8>

The payload of this node